requirements.txt
boto3==1.42.16
botocore==1.42.16
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
google-auth==2.41.1
google-auth-oauthlib==1.2.3
gspread==6.2.1
idna==3.11
iniconfig==2.3.0
jmespath==1.0.1
numpy==2.4.0
oauthlib==3.3.1
packaging==25.0
pandas==2.3.3
pluggy==1.6.0
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
s3transfer==0.16.0
six==1.17.0
tzdata==2025.3
urllib3==2.6.2
